Split column and table constraints into dedicated header files
