+++ dev/column_constraints.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "schema/constraints/primary_key.h"
+#include "schema/constraints/unique.h"
+#include "schema/constraints/auxiliary.h"
+#include "schema/constraints/default.h"
+#include "schema/constraints/collate.h"
+#include "schema/constraints/check.h"
+#include "schema/constraints/generated_always.h"
+#include "schema/constraints/null.h"
+#include "schema/constraints/unindexed.h"
