+++ dev/fts5_constraints.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "schema/constraints/prefix.h"
+#include "schema/constraints/tokenize.h"
+#include "schema/constraints/content.h"
