+++ dev/schema/column.h
@@ -15,9 +15,22 @@
 #include "../member_traits/member_traits.h"
 #include "../type_traits.h"
 #include "../type_is_nullable.h"
-#include "../constraints.h"
+#include "../column_constraints.h"
 
 namespace sqlite_orm::internal {
+    template<class T>
+    using is_column_constraint = mpl::invoke_t<mpl::disjunction<check_if<std::is_base_of, primary_key_t<>>,
+                                                                check_if_is_type<null_t>,
+                                                                check_if_is_type<not_null_t>,
+                                                                check_if_is_type<unique_t<>>,
+                                                                check_if_is_template<default_t>,
+                                                                check_if_is_template<check_t>,
+                                                                check_if_is_type<collate_constraint_t>,
+                                                                check_if<is_generated_always>,
+                                                                check_if_is_type<unindexed_t>,
+                                                                check_if<is_auxiliary>>,
+                                               T>;
+
     struct column_identifier {
 
         /**
@@ -138,37 +151,6 @@ namespace sqlite_orm::internal {
 
     template<class T>
     using is_hidden_column = polyfill::bool_constant<is_hidden_column_v<T>>;
-
-    template<class Elements>
-    using col_index_sequence_of = filter_tuple_sequence_t<Elements, is_column>;
-
-    template<class Elements, class F>
-    using col_index_sequence_with_field_type = filter_tuple_sequence_t<Elements,
-                                                                       check_if_is_type<F>::template fn,
-                                                                       field_type_t,
-                                                                       filter_tuple_sequence_t<Elements, is_column>>;
-
-    template<class Elements, template<class...> class TraitFn>
-    using col_index_sequence_with = filter_tuple_sequence_t<Elements,
-                                                            check_if_has<TraitFn>::template fn,
-                                                            constraints_type_t,
-                                                            filter_tuple_sequence_t<Elements, is_column>>;
-
-    template<class Elements, template<class...> class TraitFn>
-    using col_index_sequence_excluding = filter_tuple_sequence_t<Elements,
-                                                                 check_if_has_not<TraitFn>::template fn,
-                                                                 constraints_type_t,
-                                                                 filter_tuple_sequence_t<Elements, is_column>>;
-
-    template<class Elements>
-    using hidden_col_index_sequence_of = filter_tuple_sequence_t<Elements, is_hidden_column>;
-
-    template<class Elements, class F>
-    using all_col_index_sequence_with_field_type = filter_tuple_sequence_t<
-        Elements,
-        check_if_is_type<F>::template fn,
-        field_type_t,
-        filter_tuple_sequence_t<Elements, mpl::disjunction_fn<is_column, is_hidden_column>::template fn>>;
 
     // Custom type:
     // It is the programmer's responsibility to ensure data integrity in the value range of the custom type
