+++ dev/schema/virtual_table.h
@@ -5,17 +5,13 @@
 #include <concepts>  // std::convertible_to
 #endif
 #include <string>  //  std::string
-#include <tuple>  //  std::tuple_element, std::make_tuple
-#include <utility>  //  std::forward, std::move
+#include <utility>  //  std::move
 #endif
 
 #include "../functional/cxx_type_traits_polyfill.h"
 #include "../functional/gsl.h"
-#include "../functional/mpl.h"
 #include "../type_traits.h"
-#include "../constraints.h"
 #include "table_base.h"
-#include "column.h"
 
 namespace sqlite_orm::internal {
 #ifdef SQLITE_ORM_CPP20_CONCEPTS_SUPPORTED
