+++ dev/statement_serializer.h
@@ -37,7 +37,9 @@
 #include "ast/is_not_null.h"
 #include "core_functions.h"
 #include "window_functions.h"
-#include "constraints.h"
+#include "column_constraints.h"
+#include "table_constraints.h"
+#include "fts5_constraints.h"
 #include "conditions.h"
 #include "indexed_column.h"
 #include "function.h"
