+++ dev/storage.h
@@ -31,7 +31,8 @@
 #include "type_traits.h"
 #include "alias.h"
 #include "error_code.h"
-#include "constraints.h"
+#include "column_constraints.h"
+#include "schema/traits/constraining_traits.h"
 #include "field_printer.h"
 #include "operators.h"
 #include "select_constraints.h"
