+++ dev/table_constraints.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "schema/constraints/primary_key.h"
+#include "schema/constraints/foreign_key.h"
+#include "schema/constraints/unique.h"
+#include "schema/constraints/check.h"
