+++ include/sqlite_orm/sqlite_orm.h
@@ -3263,84 +3263,283 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
     }
 }
 
-// #include "constraints.h"
+// #include "column_constraints.h"
+
+// #include "schema/constraints/primary_key.h"
 
 #ifndef SQLITE_ORM_IMPORT_STD_MODULE
-#include <type_traits>  //  std::is_base_of, std::false_type, std::true_type
-#include <system_error>  //  std::system_error
-#include <ostream>  //  std::ostream
+#include <type_traits>  //  std::is_base_of
+#include <tuple>  //  std::tuple
+#include <utility>  //  std::forward, std::move
+#endif
+
+// #include "../../functional/cxx_type_traits_polyfill.h"
+
+// #include "../../alias_traits.h"
+
+// #include "../../field_of.h"
+
+namespace sqlite_orm::internal {
+    enum class conflict_clause_t {
+        rollback,
+        abort,
+        fail,
+        ignore,
+        replace,
+    };
+
+    struct primary_key_base {
+        enum class order_by {
+            unspecified,
+            ascending,
+            descending,
+        };
+        struct {
+            order_by asc_option = order_by::unspecified;
+            conflict_clause_t conflict_clause = conflict_clause_t::rollback;
+            bool conflict_clause_is_on = false;
+        } _options;
+    };
+
+    template<class T>
+    struct primary_key_with_autoincrement : T {
+        using primary_key_type = T;
+
+        const primary_key_type& as_base() const {
+            return *this;
+        }
+    };
+
+    /**
+     *  PRIMARY KEY constraint class.
+     *  Cs is parameter pack which contains columns (member pointers and/or function pointers). Can be empty when
+     *  used within `make_column` function.
+     */
+    template<class... Cs>
+    struct primary_key_t : primary_key_base {
+        using order_by = primary_key_base::order_by;
+        using columns_tuple = std::tuple<Cs...>;
+
+        columns_tuple _columns;
+
+        constexpr primary_key_t(columns_tuple columns) : _columns(std::move(columns)) {}
+
+        constexpr primary_key_t asc() const {
+            auto res = *this;
+            res._options.asc_option = order_by::ascending;
+            return res;
+        }
+
+        constexpr primary_key_t desc() const {
+            auto res = *this;
+            res._options.asc_option = order_by::descending;
+            return res;
+        }
+
+        constexpr primary_key_with_autoincrement<primary_key_t> autoincrement() const {
+            return {*this};
+        }
+
+        constexpr primary_key_t on_conflict_rollback() const {
+            auto res = *this;
+            res._options.conflict_clause_is_on = true;
+            res._options.conflict_clause = conflict_clause_t::rollback;
+            return res;
+        }
+
+        constexpr primary_key_t on_conflict_abort() const {
+            auto res = *this;
+            res._options.conflict_clause_is_on = true;
+            res._options.conflict_clause = conflict_clause_t::abort;
+            return res;
+        }
+
+        constexpr primary_key_t on_conflict_fail() const {
+            auto res = *this;
+            res._options.conflict_clause_is_on = true;
+            res._options.conflict_clause = conflict_clause_t::fail;
+            return res;
+        }
+
+        constexpr primary_key_t on_conflict_ignore() const {
+            auto res = *this;
+            res._options.conflict_clause_is_on = true;
+            res._options.conflict_clause = conflict_clause_t::ignore;
+            return res;
+        }
+
+        constexpr primary_key_t on_conflict_replace() const {
+            auto res = *this;
+            res._options.conflict_clause_is_on = true;
+            res._options.conflict_clause = conflict_clause_t::replace;
+            return res;
+        }
+    };
+
+    template<class T>
+    constexpr bool is_primary_key_v = std::is_base_of<primary_key_base, T>::value;
+
+    template<class T>
+    struct is_primary_key : polyfill::bool_constant<is_primary_key_v<T>> {};
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+    /**
+     *  PRIMARY KEY constraint factory function taking one or more fields, which can either be pointers to class members or column pointers.
+     */
+    template<class... Cs>
+    constexpr internal::primary_key_t<Cs...> primary_key(Cs... cs) {
+        return {{std::forward<Cs>(cs)...}};
+    }
+
+    /**
+     *  PRIMARY KEY constraint factory function taking one or more fields that are member pointers of base classes,
+     *  specifying the derived class as an explicit template argument.
+     */
+    template<class O, class... Base, class... F>
+    constexpr internal::primary_key_t<F O::*...> primary_key(F Base::*... columns) {
+        static_assert(polyfill::conjunction<internal::is_field_of<F Base::*, O>...>::value,
+                      "Fields must be from explicitly specified derived class");
+        return {{columns...}};
+    }
+
+#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
+    /**
+     *  PRIMARY KEY constraint factory function taking one or more fields that are member pointers of base classes,
+     *  specifying the derived class as an explicit template argument.
+     */
+    template<orm_table_reference auto table, class... Base, class... F>
+    constexpr auto primary_key(F Base::*... columns) {
+        return primary_key<internal::auto_decay_table_ref_t<table>>(columns...);
+    }
+#endif
+
+    /**
+     *  PRIMARY KEY column constraint factory function (used at a single column).
+     */
+    constexpr internal::primary_key_t<> primary_key() {
+        return {{}};
+    }
+}
+
+// #include "schema/constraints/unique.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
 #include <string>  //  std::string
 #include <tuple>  //  std::tuple
-#include <functional>  //  std::ref
+#include <utility>  //  std::forward, std::move
 #endif
 
-// #include "functional/cxx_type_traits_polyfill.h"
+namespace sqlite_orm::internal {
+    struct unique_base {
+        operator std::string() const {
+            return "UNIQUE";
+        }
+    };
 
-// #include "functional/mpl.h"
+    /**
+     *  UNIQUE constraint class.
+     */
+    template<class... Args>
+    struct unique_t : unique_base {
+        using columns_tuple = std::tuple<Args...>;
 
-// #include "tuple_helper/same_or_void.h"
+        columns_tuple columns;
+
+        constexpr unique_t(columns_tuple columns_) : columns(std::move(columns_)) {}
+    };
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+    /**
+     *  UNIQUE table constraint factory function.
+     */
+    template<class... Args>
+    constexpr internal::unique_t<Args...> unique(Args... args) {
+        return {{std::forward<Args>(args)...}};
+    }
+
+    /**
+     *  UNIQUE column constraint factory function.
+     */
+    constexpr internal::unique_t<> unique() {
+        return {{}};
+    }
+}
+
+// #include "schema/constraints/auxiliary.h"
 
 #ifndef SQLITE_ORM_IMPORT_STD_MODULE
-#include <type_traits>  //  std::common_type
-#ifdef SQLITE_ORM_CPP20_CONCEPTS_SUPPORTED
-#include <concepts>  //  std::same_as
+#include <type_traits>  //  std::is_same, std::false_type
 #endif
+
+namespace sqlite_orm::internal {
+#if SQLITE_VERSION_NUMBER >= 3024000
+    /**
+     *  Auxiliary virtual table column constraint
+     */
+    struct auxiliary_t {};
+
+    template<class T>
+    using is_auxiliary = std::is_same<T, auxiliary_t>;
+#else
+    template<class T>
+    using is_auxiliary = std::false_type;
+#endif
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
+#if SQLITE_VERSION_NUMBER >= 3024000
+    /**
+     *  Auxiliary virtual table column
+     */
+    constexpr internal::auxiliary_t auxiliary() {
+        return {};
+    }
+#endif
+#endif
+}
+
+// #include "schema/constraints/default.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <string>  //  std::string
+#include <utility>  //  std::move
 #endif
 
 namespace sqlite_orm::internal {
     /**
-     *  Accepts any number of arguments and evaluates a nested `type` typename as `T` if all arguments are the same, otherwise `void`.
+     *  DEFAULT constraint class.
+     *  T is a value type.
      */
-    template<class... Args>
-    struct same_or_void {
-        using type = void;
+    template<class T>
+    struct default_t {
+        using value_type = T;
+
+        value_type value;
+
+        operator std::string() const {
+            return "DEFAULT";
+        }
     };
-
-    template<class... Args>
-    using same_or_void_t = typename same_or_void<Args...>::type;
-
-#ifdef SQLITE_ORM_CPP20_CONCEPTS_SUPPORTED
-    template<class A, std::same_as<A>... Rest>
-    struct same_or_void<A, Rest...> {
-        using type = A;
-    };
-#else
-    template<class A>
-    struct same_or_void<A> {
-        using type = A;
-    };
-
-    template<class A>
-    struct same_or_void<A, A> {
-        using type = A;
-    };
-
-    template<class A, class... Args>
-    struct same_or_void<A, A, Args...> : same_or_void<A, Args...> {};
-#endif
-
-    template<class Pack>
-    struct common_type_of;
-
-    template<template<class...> class Pack, class... Types>
-    struct common_type_of<Pack<Types...>> : std::common_type<Types...> {};
-
-    /**
-     *  Accepts a pack of types and defines a nested `type` typename to a common type if possible, otherwise nonexistent.
-     *  
-     *  @note: SFINAE friendly like `std::common_type`.
-     */
-    template<class Pack>
-    using common_type_of_t = typename common_type_of<Pack>::type;
 }
 
-// #include "tuple_helper/tuple_traits.h"
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+    template<class T>
+    constexpr internal::default_t<T> default_value(T t) {
+        return {std::move(t)};
+    }
+}
 
-// #include "tuple_helper/tuple_filter.h"
+// #include "schema/constraints/collate.h"
 
-// #include "type_traits.h"
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <system_error>  //  std::system_error
+#include <string>  //  std::string
+#endif
 
-// #include "collate_argument.h"
+// #include "../../collate_argument.h"
 
 namespace sqlite_orm::internal {
     enum class collate_argument {
@@ -3350,13 +3549,181 @@ namespace sqlite_orm::internal {
     };
 }
 
-// #include "error_code.h"
+// #include "../../error_code.h"
 
-// #include "field_of.h"
+namespace sqlite_orm::internal {
+    struct collate_constraint_t {
+        collate_argument argument = collate_argument::binary;
 
-// #include "table_type_of.h"
+        static std::string string_from_collate_argument(collate_argument argument) {
+            switch (argument) {
+                case collate_argument::binary:
+                    return "BINARY";
+                case collate_argument::nocase:
+                    return "NOCASE";
+                case collate_argument::rtrim:
+                    return "RTRIM";
+            }
+            throw std::system_error{orm_error_code::invalid_collate_argument_enum};
+        }
+    };
+}
 
-// #include "type_printer.h"
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+    constexpr internal::collate_constraint_t collate_nocase() {
+        return {internal::collate_argument::nocase};
+    }
+
+    constexpr internal::collate_constraint_t collate_binary() {
+        return {internal::collate_argument::binary};
+    }
+
+    constexpr internal::collate_constraint_t collate_rtrim() {
+        return {internal::collate_argument::rtrim};
+    }
+}
+
+// #include "schema/constraints/check.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <utility>  //  std::move
+#endif
+
+namespace sqlite_orm::internal {
+    template<class T>
+    struct check_t {
+        using expression_type = T;
+
+        expression_type expression;
+    };
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+    template<class T>
+    constexpr internal::check_t<T> check(T t) {
+        return {std::move(t)};
+    }
+}
+
+// #include "schema/constraints/generated_always.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <utility>  //  std::move
+#endif
+
+// #include "../../functional/cxx_type_traits_polyfill.h"
+
+namespace sqlite_orm::internal {
+    struct basic_generated_always {
+        enum class storage_type {
+            not_specified,
+            virtual_,
+            stored,
+        };
+
+#if SQLITE_VERSION_NUMBER >= 3031000
+        bool _full = true;
+        storage_type _storage = storage_type::not_specified;
+#endif
+    };
+
+#if SQLITE_VERSION_NUMBER >= 3031000
+    template<class T>
+    struct generated_always_t : basic_generated_always {
+        using expression_type = T;
+
+        expression_type _expression;
+
+        constexpr generated_always_t<T> virtual_() && {
+            return {_full, storage_type::virtual_, std::move(_expression)};
+        }
+
+        constexpr generated_always_t<T> stored() && {
+            return {_full, storage_type::stored, std::move(_expression)};
+        }
+    };
+#endif
+
+    template<class T>
+    constexpr bool is_generated_always_v =
+#if SQLITE_VERSION_NUMBER >= 3031000
+        polyfill::is_specialization_of<T, generated_always_t>::value;
+#else
+        false;
+#endif
+
+    template<class T>
+    struct is_generated_always : polyfill::bool_constant<is_generated_always_v<T>> {};
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3031000
+    template<class T>
+    constexpr internal::generated_always_t<T> generated_always_as(T expression) {
+        return {true, internal::basic_generated_always::storage_type::not_specified, std::move(expression)};
+    }
+
+    template<class T>
+    constexpr internal::generated_always_t<T> as(T expression) {
+        return {false, internal::basic_generated_always::storage_type::not_specified, std::move(expression)};
+    }
+#endif
+}
+
+// #include "schema/constraints/null.h"
+
+namespace sqlite_orm::internal {
+    struct null_t {};
+
+    struct not_null_t {};
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+    constexpr internal::null_t null() {
+        return {};
+    }
+
+    constexpr internal::not_null_t not_null() {
+        return {};
+    }
+}
+
+// #include "schema/constraints/unindexed.h"
+
+namespace sqlite_orm::internal {
+    struct unindexed_t {};
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
+    /**
+     *  UNINDEXED column constraint factory function. Used in FTS virtual tables.
+     *
+     *  https://www.sqlite.org/fts5.html#the_unindexed_column_option
+     */
+    constexpr internal::unindexed_t unindexed() {
+        return {};
+    }
+#endif
+}
+
+// #include "schema/traits/constraining_traits.h"
+
+// #include "../../functional/cxx_type_traits_polyfill.h"
+
+// #include "../../functional/mpl.h"
+
+// #include "../../type_traits.h"
+
+// #include "../constraints/default.h"
+
+// #include "is_rowid_alias_capable.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <type_traits>  //  std::is_base_of, std::is_integral, std::is_signed, std::enable_if
+#endif
+
+// #include "../../type_printer.h"
 
 #ifndef SQLITE_ORM_IMPORT_STD_MODULE
 #include <string>  //  std::string
@@ -3478,491 +3845,15 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 }
 
 namespace sqlite_orm::internal {
-    enum class conflict_clause_t {
-        rollback,
-        abort,
-        fail,
-        ignore,
-        replace,
-    };
-
-    struct primary_key_base {
-        enum class order_by {
-            unspecified,
-            ascending,
-            descending,
-        };
-        struct {
-            order_by asc_option = order_by::unspecified;
-            conflict_clause_t conflict_clause = conflict_clause_t::rollback;
-            bool conflict_clause_is_on = false;
-        } _options;
-    };
-
-    template<class T>
-    struct primary_key_with_autoincrement : T {
-        using primary_key_type = T;
-
-        const primary_key_type& as_base() const {
-            return *this;
-        }
-    };
-
-    /**
-     *  PRIMARY KEY constraint class.
-     *  Cs is parameter pack which contains columns (member pointers and/or function pointers). Can be empty when
-     *  used within `make_column` function.
-     */
-    template<class... Cs>
-    struct primary_key_t : primary_key_base {
-        using order_by = primary_key_base::order_by;
-        using columns_tuple = std::tuple<Cs...>;
-
-        columns_tuple _columns;
-
-        constexpr primary_key_t(columns_tuple columns) : _columns(std::move(columns)) {}
-
-        constexpr primary_key_t asc() const {
-            auto res = *this;
-            res._options.asc_option = order_by::ascending;
-            return res;
-        }
-
-        constexpr primary_key_t desc() const {
-            auto res = *this;
-            res._options.asc_option = order_by::descending;
-            return res;
-        }
-
-        constexpr primary_key_with_autoincrement<primary_key_t> autoincrement() const {
-            return {*this};
-        }
-
-        constexpr primary_key_t on_conflict_rollback() const {
-            auto res = *this;
-            res._options.conflict_clause_is_on = true;
-            res._options.conflict_clause = conflict_clause_t::rollback;
-            return res;
-        }
-
-        constexpr primary_key_t on_conflict_abort() const {
-            auto res = *this;
-            res._options.conflict_clause_is_on = true;
-            res._options.conflict_clause = conflict_clause_t::abort;
-            return res;
-        }
-
-        constexpr primary_key_t on_conflict_fail() const {
-            auto res = *this;
-            res._options.conflict_clause_is_on = true;
-            res._options.conflict_clause = conflict_clause_t::fail;
-            return res;
-        }
-
-        constexpr primary_key_t on_conflict_ignore() const {
-            auto res = *this;
-            res._options.conflict_clause_is_on = true;
-            res._options.conflict_clause = conflict_clause_t::ignore;
-            return res;
-        }
-
-        constexpr primary_key_t on_conflict_replace() const {
-            auto res = *this;
-            res._options.conflict_clause_is_on = true;
-            res._options.conflict_clause = conflict_clause_t::replace;
-            return res;
-        }
-    };
-
-    struct unique_base {
-        operator std::string() const {
-            return "UNIQUE";
-        }
-    };
-
-    /**
-     *  UNIQUE constraint class.
-     */
-    template<class... Args>
-    struct unique_t : unique_base {
-        using columns_tuple = std::tuple<Args...>;
-
-        columns_tuple columns;
-
-        constexpr unique_t(columns_tuple columns_) : columns(std::move(columns_)) {}
-    };
-
-    struct unindexed_t {};
-
-    template<class T>
-    struct prefix_t {
-        using value_type = T;
-
-        value_type value;
-    };
-
-    template<class T>
-    struct tokenize_t {
-        using value_type = T;
-
-        value_type value;
-    };
-
-    template<class T>
-    struct content_t {
-        using value_type = T;
-
-        value_type value;
-    };
-
-    template<class T>
-    struct table_content_t {
-        using mapped_type = T;
-    };
-
-#if SQLITE_VERSION_NUMBER >= 3024000
-    /** 
-     *  Auxiliary virtual table column constraint
-     */
-    struct auxiliary_t {};
-
-    template<class T>
-    using is_auxiliary = std::is_same<T, auxiliary_t>;
-#else
-    template<class T>
-    using is_auxiliary = std::false_type;
-#endif
-
-    /**
-     *  DEFAULT constraint class.
-     *  T is a value type.
-     */
-    template<class T>
-    struct default_t {
-        using value_type = T;
-
-        value_type value;
-
-        operator std::string() const {
-            return "DEFAULT";
-        }
-    };
-
-#if SQLITE_VERSION_NUMBER >= 3006019
-    /**
-     *  FOREIGN KEY constraint class.
-     *  Cs are columns which has foreign key
-     *  Rs are column which C references to
-     *  Available since SQLite 3.6.19
-     */
-    template<class A, class B>
-    struct foreign_key_t;
-
-    enum class foreign_key_action {
-        none,  //  not specified
-        no_action,
-        restrict_,
-        set_null,
-        set_default,
-        cascade,
-    };
-
-    inline std::ostream& operator<<(std::ostream& os, foreign_key_action action) {
-        switch (action) {
-            case foreign_key_action::no_action:
-                os << "NO ACTION";
-                break;
-            case foreign_key_action::restrict_:
-                os << "RESTRICT";
-                break;
-            case foreign_key_action::set_null:
-                os << "SET NULL";
-                break;
-            case foreign_key_action::set_default:
-                os << "SET DEFAULT";
-                break;
-            case foreign_key_action::cascade:
-                os << "CASCADE";
-                break;
-            case foreign_key_action::none:
-                break;
-        }
-        return os;
-    }
-
-    struct on_update_delete_base {
-        const bool update;  //  true if update and false if delete
-
-        operator std::string() const {
-            if (this->update) {
-                return "ON UPDATE";
-            } else {
-                return "ON DELETE";
-            }
-        }
-    };
-
-    /**
-     *  F - foreign key class
-     */
-    template<class F>
-    struct on_update_delete_t : on_update_delete_base {
-        using foreign_key_type = F;
-
-        const foreign_key_type& fk;
-
-        foreign_key_action _action = foreign_key_action::none;
-
-        on_update_delete_t(const on_update_delete_t&) = delete;
-        on_update_delete_t& operator=(const on_update_delete_t&) = delete;
-
-        on_update_delete_t(foreign_key_type& fk_, decltype(update) update_, foreign_key_action action_) :
-            on_update_delete_base{update_}, fk(fk_), _action(action_) {}
-
-        foreign_key_type no_action() const {
-            auto res = this->fk;
-            if (update) {
-                res.on_update._action = foreign_key_action::no_action;
-            } else {
-                res.on_delete._action = foreign_key_action::no_action;
-            }
-            return res;
-        }
-
-        foreign_key_type restrict_() const {
-            auto res = this->fk;
-            if (update) {
-                res.on_update._action = foreign_key_action::restrict_;
-            } else {
-                res.on_delete._action = foreign_key_action::restrict_;
-            }
-            return res;
-        }
-
-        foreign_key_type set_null() const {
-            auto res = this->fk;
-            if (update) {
-                res.on_update._action = foreign_key_action::set_null;
-            } else {
-                res.on_delete._action = foreign_key_action::set_null;
-            }
-            return res;
-        }
-
-        foreign_key_type set_default() const {
-            auto res = this->fk;
-            if (update) {
-                res.on_update._action = foreign_key_action::set_default;
-            } else {
-                res.on_delete._action = foreign_key_action::set_default;
-            }
-            return res;
-        }
-
-        foreign_key_type cascade() const {
-            auto res = this->fk;
-            if (update) {
-                res.on_update._action = foreign_key_action::cascade;
-            } else {
-                res.on_delete._action = foreign_key_action::cascade;
-            }
-            return res;
-        }
-
-        explicit operator bool() const {
-            return _action != foreign_key_action::none;
-        }
-    };
-
-    template<class F>
-    bool operator==(const on_update_delete_t<F>& lhs, const on_update_delete_t<F>& rhs) {
-        return lhs._action == rhs._action;
-    }
-
-    template<class... Cs, class... Rs>
-    struct foreign_key_t<std::tuple<Cs...>, std::tuple<Rs...>> {
-        using columns_type = std::tuple<Cs...>;
-        using references_type = std::tuple<Rs...>;
-
-        /**
-         *  Holds obect type of all referenced columns.
-         */
-        using target_type = same_or_void_t<table_type_of_t<Rs>...>;
-
-        /**
-         *  Holds obect type of all source columns.
-         */
-        using source_type = same_or_void_t<table_type_of_t<Cs>...>;
-
-        columns_type _columns;
-        references_type _references;
-
-        on_update_delete_t<foreign_key_t> on_update;
-        on_update_delete_t<foreign_key_t> on_delete;
-
-        static_assert(std::tuple_size<columns_type>::value == std::tuple_size<references_type>::value,
-                      "Columns size must be equal to references tuple");
-        static_assert(!std::is_same<source_type, void>::value, "All columns must have the same mapped type");
-        static_assert(!std::is_same<target_type, void>::value, "All references must have the same mapped type");
-
-        foreign_key_t(columns_type columns, references_type references) :
-            _columns(std::move(columns)), _references(std::move(references)),
-            on_update(std::ref(*this), true, foreign_key_action::none),
-            on_delete(std::ref(*this), false, foreign_key_action::none) {}
-
-        foreign_key_t(const foreign_key_t& other) :
-            _columns(other._columns), _references(other._references),
-            on_update(std::ref(*this), true, other.on_update._action),
-            on_delete(std::ref(*this), false, other.on_delete._action) {}
-
-        foreign_key_t& operator=(const foreign_key_t&) = delete;
-
-#ifdef SQLITE_ORM_DEFAULT_COMPARISONS_SUPPORTED
-        friend bool operator==(const foreign_key_t& lhs, const foreign_key_t& rhs) = default;
-#else
-        friend bool operator==(const foreign_key_t& lhs, const foreign_key_t& rhs) {
-            return lhs._columns == rhs._columns && lhs._references == rhs._references &&
-                   lhs.on_update == rhs.on_update && lhs.on_delete == rhs.on_delete;
-        }
-#endif
-    };
-
-    /**
-     *  Cs can be a class member pointer or column pointer
-     *  Available since SQLite 3.6.19
-     */
-    template<class... Cs>
-    struct foreign_key_intermediate_t {
-        using tuple_type = std::tuple<Cs...>;
-
-        tuple_type _columns;
-
-        /**
-         *  Specify one or more target fields, which can either be pointers to class members or column pointers.
-         */
-        template<class... Rs>
-        foreign_key_t<tuple_type, std::tuple<Rs...>> references(Rs... refs) {
-            return {std::move(_columns), {std::forward<Rs>(refs)...}};
-        }
-
-        /**
-         *  Specify one or more target fields that are member pointers of base classes,
-         *  specifying the derived class as an explicit template argument.
-         */
-        template<class O, class... Base, class... F>
-        foreign_key_t<tuple_type, std::tuple<F O::*...>> references(F Base::*... refs) {
-            static_assert(polyfill::conjunction<is_field_of<F Base::*, O>...>::value,
-                          "Referenced fields must be from explicitly specified derived class");
-            return {std::move(_columns), {refs...}};
-        }
-
-#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
-        /**
-         *  Specify one or more target fields that are member pointers of base classes,
-         *  specifying the derived class as an explicit template argument.
-         */
-        template<orm_table_reference auto table, class... Base, class... F>
-        auto references(F Base::*... refs) {
-            return this->references<auto_decay_table_ref_t<table>>(refs...);
-        }
-#endif
-    };
-#endif
-
-    struct collate_constraint_t {
-        collate_argument argument = collate_argument::binary;
-
-        static std::string string_from_collate_argument(collate_argument argument) {
-            switch (argument) {
-                case collate_argument::binary:
-                    return "BINARY";
-                case collate_argument::nocase:
-                    return "NOCASE";
-                case collate_argument::rtrim:
-                    return "RTRIM";
-            }
-            throw std::system_error{orm_error_code::invalid_collate_argument_enum};
-        }
-    };
-
-    template<class T>
-    struct check_t {
-        using expression_type = T;
-
-        expression_type expression;
-    };
-
-    struct basic_generated_always {
-        enum class storage_type {
-            not_specified,
-            virtual_,
-            stored,
-        };
-
-#if SQLITE_VERSION_NUMBER >= 3031000
-        bool _full = true;
-        storage_type _storage = storage_type::not_specified;
-#endif
-    };
-
-#if SQLITE_VERSION_NUMBER >= 3031000
-    template<class T>
-    struct generated_always_t : basic_generated_always {
-        using expression_type = T;
-
-        expression_type _expression;
-
-        constexpr generated_always_t<T> virtual_() && {
-            return {_full, storage_type::virtual_, std::move(_expression)};
-        }
-
-        constexpr generated_always_t<T> stored() && {
-            return {_full, storage_type::stored, std::move(_expression)};
-        }
-    };
-#endif
-
-    struct null_t {};
-
-    struct not_null_t {};
-
-    template<class T>
-    inline constexpr bool is_foreign_key_v =
-#if SQLITE_VERSION_NUMBER >= 3006019
-        polyfill::is_specialization_of<T, foreign_key_t>::value;
-#else
-        false;
-#endif
-
-    template<class T>
-    struct is_foreign_key : polyfill::bool_constant<is_foreign_key_v<T>> {};
-
-    template<class T>
-    inline constexpr bool is_primary_key_v = std::is_base_of<primary_key_base, T>::value;
-
-    template<class T>
-    struct is_primary_key : polyfill::bool_constant<is_primary_key_v<T>> {};
-
-    template<class T>
-    inline constexpr bool is_generated_always_v =
-#if SQLITE_VERSION_NUMBER >= 3031000
-        polyfill::is_specialization_of<T, generated_always_t>::value;
-#else
-        false;
-#endif
-
-    template<class T>
-    struct is_generated_always : polyfill::bool_constant<is_generated_always_v<T>> {};
-
     // Custom type:
     // It is the programmer's responsibility to ensure data integrity in the value range of the custom type
     // and in purview of SQLite using a 64-bit signed integer.
     template<class F, class SFINAE = void>
-    inline constexpr bool is_rowid_alias_capable_v = std::is_base_of<integer_printer, type_printer<F>>::value;
+    constexpr bool is_rowid_alias_capable_v = std::is_base_of<integer_printer, type_printer<F>>::value;
 
     // For 64-bit signed integer type: capable
     template<class F>
-    inline constexpr bool
+    constexpr bool
         is_rowid_alias_capable_v<F,
                                  std::enable_if_t<std::is_integral<F>::value &&
                                                   (sizeof(F) == sizeof(sqlite_int64) &&
@@ -3973,7 +3864,7 @@ namespace sqlite_orm::internal {
     // It is the programmer's responsibility to ensure data integrity in the value range of the integral type
     // and in purview of SQLite using a 64-bit signed integer.
     template<class F>
-    inline constexpr bool
+    constexpr bool
         is_rowid_alias_capable_v<F,
                                  std::enable_if_t<std::is_integral<F>::value &&
                                                   (sizeof(F) != sizeof(sqlite_int64) ||
@@ -3982,20 +3873,22 @@ namespace sqlite_orm::internal {
 
     // For non-integer types: unsuitable
     template<class F>
-    inline constexpr bool
+    constexpr bool
         is_rowid_alias_capable_v<F, std::enable_if_t<!std::is_base_of<integer_printer, type_printer<F>>::value>> =
             false;
+}
 
-    /** 
+namespace sqlite_orm::internal {
+    /**
      *  COLUMN PRIMARY KEY INSERTABLE traits.
-     *  
+     *
      *  A column primary key is considered implicitly insertable if:
      *  - it is an INTEGER PRIMARY KEY (and thus an alias for the "rowid" key),
      *  - or has a default value.
-     *  
+     *
      *  In terms of C++ types, this means that the field type must be capable of representing a 64-bit signed integer,
      *  or the column is declared with a DEFAULT constraint.
-     *  
+     *
      *  Implementation note: using a struct template in favor of a template alias so that the stack leading to a deprecation message is shorter.
      */
     template<typename Column>
@@ -4004,222 +3897,6 @@ namespace sqlite_orm::internal {
               mpl::disjunction<mpl::always<polyfill::bool_constant<is_rowid_alias_capable_v<field_type_t<Column>>>>,
                                check_if_has_template<default_t>>,
               constraints_type_t<Column>> {};
-
-    template<class T>
-    using is_column_constraint = mpl::invoke_t<mpl::disjunction<check_if<std::is_base_of, primary_key_t<>>,
-                                                                check_if_is_type<null_t>,
-                                                                check_if_is_type<not_null_t>,
-                                                                check_if_is_type<unique_t<>>,
-                                                                check_if_is_template<default_t>,
-                                                                check_if_is_template<check_t>,
-                                                                check_if_is_type<collate_constraint_t>,
-                                                                check_if<is_generated_always>,
-                                                                check_if_is_type<unindexed_t>,
-                                                                check_if<is_auxiliary>>,
-                                               T>;
-}
-
-SQLITE_ORM_EXPORT namespace sqlite_orm {
-#if SQLITE_VERSION_NUMBER >= 3031000
-    template<class T>
-    constexpr internal::generated_always_t<T> generated_always_as(T expression) {
-        return {true, internal::basic_generated_always::storage_type::not_specified, std::move(expression)};
-    }
-
-    template<class T>
-    constexpr internal::generated_always_t<T> as(T expression) {
-        return {false, internal::basic_generated_always::storage_type::not_specified, std::move(expression)};
-    }
-#endif
-
-#if SQLITE_VERSION_NUMBER >= 3006019
-    /**
-     *  FOREIGN KEY constraint factory function taking one or more fields, which can either be pointers to class members or column pointers.
-     *  Available since SQLite 3.6.19
-     */
-    template<class... Cs>
-    constexpr internal::foreign_key_intermediate_t<Cs...> foreign_key(Cs... columns) {
-        return {{std::forward<Cs>(columns)...}};
-    }
-
-    /**
-     *  FOREIGN KEY constraint factory function taking one or more fields that are member pointers of base classes,
-     *  specifying the derived class as an explicit template argument.
-     *  Available since SQLite 3.6.19
-     */
-    template<class O, class... Base, class... F>
-    constexpr internal::foreign_key_intermediate_t<F O::*...> foreign_key(F Base::*... columns) {
-        static_assert(polyfill::conjunction<internal::is_field_of<F Base::*, O>...>::value,
-                      "Fields must be from explicitly specified derived class");
-        return {{columns...}};
-    }
-
-#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
-    /**
-     *  FOREIGN KEY constraint factory function taking one or more fields that are member pointers of base classes,
-     *  specifying the derived class as an explicit template argument.
-     *  Available since SQLite 3.6.19
-     */
-    template<orm_table_reference auto table, class... Base, class... F>
-    constexpr auto foreign_key(F Base::*... columns) {
-        return foreign_key<internal::auto_decay_table_ref_t<table>>(columns...);
-    }
-#endif
-#endif
-
-    /**
-     *  UNIQUE table constraint factory function.
-     */
-    template<class... Args>
-    constexpr internal::unique_t<Args...> unique(Args... args) {
-        return {{std::forward<Args>(args)...}};
-    }
-
-    /**
-     *  UNIQUE column constraint factory function.
-     */
-    constexpr internal::unique_t<> unique() {
-        return {{}};
-    }
-
-#if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
-    /**
-     *  UNINDEXED column constraint factory function. Used in FTS virtual tables.
-     *  
-     *  https://www.sqlite.org/fts5.html#the_unindexed_column_option
-     */
-    constexpr internal::unindexed_t unindexed() {
-        return {};
-    }
-
-    /**
-     *  prefix=N table constraint factory function. Used in FTS virtual tables.
-     *  
-     *  https://www.sqlite.org/fts5.html#prefix_indexes
-     */
-    template<class T>
-    constexpr internal::prefix_t<T> prefix(T value) {
-        return {std::move(value)};
-    }
-
-    /**
-     *  tokenize='...'' table constraint factory function. Used in FTS virtual tables.
-     *  
-     *  https://www.sqlite.org/fts5.html#tokenizers
-     */
-    template<class T>
-    constexpr internal::tokenize_t<T> tokenize(T value) {
-        return {std::move(value)};
-    }
-
-    /**
-     *  content='' table constraint factory function. Used in FTS virtual tables.
-     *  
-     *  https://www.sqlite.org/fts5.html#contentless_tables
-     */
-    template<class T>
-    constexpr internal::content_t<T> content(T value) {
-        return {std::move(value)};
-    }
-
-    /**
-     *  content='table' table constraint factory function. Used in FTS virtual tables.
-     *  
-     *  https://www.sqlite.org/fts5.html#external_content_tables
-     */
-    template<class T>
-    constexpr internal::table_content_t<T> content() {
-        return {};
-    }
-
-#if SQLITE_VERSION_NUMBER >= 3024000
-    /** 
-     *  Auxiliary virtual table column
-     */
-    constexpr inline internal::auxiliary_t auxiliary() {
-        return {};
-    }
-#endif
-
-#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
-    /**
-     *  content='table' table constraint factory function. Used in FTS virtual tables.
-     *  
-     *  https://www.sqlite.org/fts5.html#external_content_tables
-     */
-    template<orm_table_reference auto table>
-    constexpr auto content() {
-        return content<internal::auto_decay_table_ref_t<table>>();
-    }
-#endif
-#endif
-
-    /**
-     *  PRIMARY KEY constraint factory function taking one or more fields, which can either be pointers to class members or column pointers.
-     */
-    template<class... Cs>
-    constexpr internal::primary_key_t<Cs...> primary_key(Cs... cs) {
-        return {{std::forward<Cs>(cs)...}};
-    }
-
-    /**
-     *  PRIMARY KEY constraint factory function taking one or more fields that are member pointers of base classes,
-     *  specifying the derived class as an explicit template argument.
-     */
-    template<class O, class... Base, class... F>
-    constexpr internal::primary_key_t<F O::*...> primary_key(F Base::*... columns) {
-        static_assert(polyfill::conjunction<internal::is_field_of<F Base::*, O>...>::value,
-                      "Fields must be from explicitly specified derived class");
-        return {{columns...}};
-    }
-
-#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
-    /**
-     *  PRIMARY KEY constraint factory function taking one or more fields that are member pointers of base classes,
-     *  specifying the derived class as an explicit template argument.
-     */
-    template<orm_table_reference auto table, class... Base, class... F>
-    constexpr auto primary_key(F Base::*... columns) {
-        return primary_key<internal::auto_decay_table_ref_t<table>>(columns...);
-    }
-#endif
-
-    /**
-     *  PRIMARY KEY column constraint factory function (used at a single column).
-     */
-    constexpr internal::primary_key_t<> primary_key() {
-        return {{}};
-    }
-
-    template<class T>
-    constexpr internal::default_t<T> default_value(T t) {
-        return {std::move(t)};
-    }
-
-    constexpr internal::collate_constraint_t collate_nocase() {
-        return {internal::collate_argument::nocase};
-    }
-
-    constexpr internal::collate_constraint_t collate_binary() {
-        return {internal::collate_argument::binary};
-    }
-
-    constexpr internal::collate_constraint_t collate_rtrim() {
-        return {internal::collate_argument::rtrim};
-    }
-
-    template<class T>
-    constexpr internal::check_t<T> check(T t) {
-        return {std::move(t)};
-    }
-
-    constexpr internal::null_t null() {
-        return {};
-    }
-
-    constexpr internal::not_null_t not_null() {
-        return {};
-    }
 }
 
 // #include "field_printer.h"
@@ -5058,7 +4735,7 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 
 // #include "collate_argument.h"
 
-// #include "constraints.h"
+// #include "schema/constraints/collate.h"
 
 // #include "optional_container.h"
 
@@ -8986,9 +8663,22 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
     };
 }
 
-// #include "../constraints.h"
+// #include "../column_constraints.h"
 
 namespace sqlite_orm::internal {
+    template<class T>
+    using is_column_constraint = mpl::invoke_t<mpl::disjunction<check_if<std::is_base_of, primary_key_t<>>,
+                                                                check_if_is_type<null_t>,
+                                                                check_if_is_type<not_null_t>,
+                                                                check_if_is_type<unique_t<>>,
+                                                                check_if_is_template<default_t>,
+                                                                check_if_is_template<check_t>,
+                                                                check_if_is_type<collate_constraint_t>,
+                                                                check_if<is_generated_always>,
+                                                                check_if_is_type<unindexed_t>,
+                                                                check_if<is_auxiliary>>,
+                                               T>;
+
     struct column_identifier {
 
         /**
@@ -9109,37 +8799,6 @@ namespace sqlite_orm::internal {
 
     template<class T>
     using is_hidden_column = polyfill::bool_constant<is_hidden_column_v<T>>;
-
-    template<class Elements>
-    using col_index_sequence_of = filter_tuple_sequence_t<Elements, is_column>;
-
-    template<class Elements, class F>
-    using col_index_sequence_with_field_type = filter_tuple_sequence_t<Elements,
-                                                                       check_if_is_type<F>::template fn,
-                                                                       field_type_t,
-                                                                       filter_tuple_sequence_t<Elements, is_column>>;
-
-    template<class Elements, template<class...> class TraitFn>
-    using col_index_sequence_with = filter_tuple_sequence_t<Elements,
-                                                            check_if_has<TraitFn>::template fn,
-                                                            constraints_type_t,
-                                                            filter_tuple_sequence_t<Elements, is_column>>;
-
-    template<class Elements, template<class...> class TraitFn>
-    using col_index_sequence_excluding = filter_tuple_sequence_t<Elements,
-                                                                 check_if_has_not<TraitFn>::template fn,
-                                                                 constraints_type_t,
-                                                                 filter_tuple_sequence_t<Elements, is_column>>;
-
-    template<class Elements>
-    using hidden_col_index_sequence_of = filter_tuple_sequence_t<Elements, is_hidden_column>;
-
-    template<class Elements, class F>
-    using all_col_index_sequence_with_field_type = filter_tuple_sequence_t<
-        Elements,
-        check_if_is_type<F>::template fn,
-        field_type_t,
-        filter_tuple_sequence_t<Elements, mpl::disjunction_fn<is_column, is_hidden_column>::template fn>>;
 
     // Custom type:
     // It is the programmer's responsibility to ensure data integrity in the value range of the custom type
@@ -11058,6 +10717,60 @@ namespace sqlite_orm::internal {
 // #include "tuple_helper/tuple_transformer.h"
 
 // #include "tuple_helper/same_or_void.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <type_traits>  //  std::common_type
+#ifdef SQLITE_ORM_CPP20_CONCEPTS_SUPPORTED
+#include <concepts>  //  std::same_as
+#endif
+#endif
+
+namespace sqlite_orm::internal {
+    /**
+     *  Accepts any number of arguments and evaluates a nested `type` typename as `T` if all arguments are the same, otherwise `void`.
+     */
+    template<class... Args>
+    struct same_or_void {
+        using type = void;
+    };
+
+    template<class... Args>
+    using same_or_void_t = typename same_or_void<Args...>::type;
+
+#ifdef SQLITE_ORM_CPP20_CONCEPTS_SUPPORTED
+    template<class A, std::same_as<A>... Rest>
+    struct same_or_void<A, Rest...> {
+        using type = A;
+    };
+#else
+    template<class A>
+    struct same_or_void<A> {
+        using type = A;
+    };
+
+    template<class A>
+    struct same_or_void<A, A> {
+        using type = A;
+    };
+
+    template<class A, class... Args>
+    struct same_or_void<A, A, Args...> : same_or_void<A, Args...> {};
+#endif
+
+    template<class Pack>
+    struct common_type_of;
+
+    template<template<class...> class Pack, class... Types>
+    struct common_type_of<Pack<Types...>> : std::common_type<Types...> {};
+
+    /**
+     *  Accepts a pack of types and defines a nested `type` typename to a common type if possible, otherwise nonexistent.
+     *  
+     *  @note: SFINAE friendly like `std::common_type`.
+     */
+    template<class Pack>
+    using common_type_of_t = typename common_type_of<Pack>::type;
+}
 
 // #include "type_traits.h"
 
@@ -13233,7 +12946,305 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 
 // #include "../type_traits.h"
 
-// #include "../constraints.h"
+// #include "../table_constraints.h"
+
+// #include "schema/constraints/primary_key.h"
+
+// #include "schema/constraints/foreign_key.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <type_traits>  //  std::is_same
+#include <ostream>  //  std::ostream
+#include <string>  //  std::string
+#include <tuple>  //  std::tuple, std::tuple_size
+#include <functional>  //  std::ref
+#include <utility>  //  std::forward, std::move
+#endif
+
+// #include "../../functional/cxx_type_traits_polyfill.h"
+
+// #include "../../tuple_helper/same_or_void.h"
+
+// #include "../../alias_traits.h"
+
+// #include "../../field_of.h"
+
+// #include "../../table_type_of.h"
+
+namespace sqlite_orm::internal {
+#if SQLITE_VERSION_NUMBER >= 3006019
+    /**
+     *  FOREIGN KEY constraint class.
+     *  Cs are columns which has foreign key
+     *  Rs are column which C references to
+     *  Available since SQLite 3.6.19
+     */
+    template<class A, class B>
+    struct foreign_key_t;
+
+    enum class foreign_key_action {
+        none,  //  not specified
+        no_action,
+        restrict_,
+        set_null,
+        set_default,
+        cascade,
+    };
+
+    inline std::ostream& operator<<(std::ostream& os, foreign_key_action action) {
+        switch (action) {
+            case foreign_key_action::no_action:
+                os << "NO ACTION";
+                break;
+            case foreign_key_action::restrict_:
+                os << "RESTRICT";
+                break;
+            case foreign_key_action::set_null:
+                os << "SET NULL";
+                break;
+            case foreign_key_action::set_default:
+                os << "SET DEFAULT";
+                break;
+            case foreign_key_action::cascade:
+                os << "CASCADE";
+                break;
+            case foreign_key_action::none:
+                break;
+        }
+        return os;
+    }
+
+    struct on_update_delete_base {
+        const bool update;  //  true if update and false if delete
+
+        operator std::string() const {
+            if (this->update) {
+                return "ON UPDATE";
+            } else {
+                return "ON DELETE";
+            }
+        }
+    };
+
+    /**
+     *  F - foreign key class
+     */
+    template<class F>
+    struct on_update_delete_t : on_update_delete_base {
+        using foreign_key_type = F;
+
+        const foreign_key_type& fk;
+
+        foreign_key_action _action = foreign_key_action::none;
+
+        on_update_delete_t(const on_update_delete_t&) = delete;
+        on_update_delete_t& operator=(const on_update_delete_t&) = delete;
+
+        on_update_delete_t(foreign_key_type& fk_, decltype(update) update_, foreign_key_action action_) :
+            on_update_delete_base{update_}, fk(fk_), _action(action_) {}
+
+        foreign_key_type no_action() const {
+            auto res = this->fk;
+            if (update) {
+                res.on_update._action = foreign_key_action::no_action;
+            } else {
+                res.on_delete._action = foreign_key_action::no_action;
+            }
+            return res;
+        }
+
+        foreign_key_type restrict_() const {
+            auto res = this->fk;
+            if (update) {
+                res.on_update._action = foreign_key_action::restrict_;
+            } else {
+                res.on_delete._action = foreign_key_action::restrict_;
+            }
+            return res;
+        }
+
+        foreign_key_type set_null() const {
+            auto res = this->fk;
+            if (update) {
+                res.on_update._action = foreign_key_action::set_null;
+            } else {
+                res.on_delete._action = foreign_key_action::set_null;
+            }
+            return res;
+        }
+
+        foreign_key_type set_default() const {
+            auto res = this->fk;
+            if (update) {
+                res.on_update._action = foreign_key_action::set_default;
+            } else {
+                res.on_delete._action = foreign_key_action::set_default;
+            }
+            return res;
+        }
+
+        foreign_key_type cascade() const {
+            auto res = this->fk;
+            if (update) {
+                res.on_update._action = foreign_key_action::cascade;
+            } else {
+                res.on_delete._action = foreign_key_action::cascade;
+            }
+            return res;
+        }
+
+        explicit operator bool() const {
+            return _action != foreign_key_action::none;
+        }
+    };
+
+    template<class F>
+    bool operator==(const on_update_delete_t<F>& lhs, const on_update_delete_t<F>& rhs) {
+        return lhs._action == rhs._action;
+    }
+
+    template<class... Cs, class... Rs>
+    struct foreign_key_t<std::tuple<Cs...>, std::tuple<Rs...>> {
+        using columns_type = std::tuple<Cs...>;
+        using references_type = std::tuple<Rs...>;
+
+        /**
+         *  Holds obect type of all referenced columns.
+         */
+        using target_type = same_or_void_t<table_type_of_t<Rs>...>;
+
+        /**
+         *  Holds obect type of all source columns.
+         */
+        using source_type = same_or_void_t<table_type_of_t<Cs>...>;
+
+        columns_type _columns;
+        references_type _references;
+
+        on_update_delete_t<foreign_key_t> on_update;
+        on_update_delete_t<foreign_key_t> on_delete;
+
+        static_assert(std::tuple_size<columns_type>::value == std::tuple_size<references_type>::value,
+                      "Columns size must be equal to references tuple");
+        static_assert(!std::is_same<source_type, void>::value, "All columns must have the same mapped type");
+        static_assert(!std::is_same<target_type, void>::value, "All references must have the same mapped type");
+
+        foreign_key_t(columns_type columns, references_type references) :
+            _columns(std::move(columns)), _references(std::move(references)),
+            on_update(std::ref(*this), true, foreign_key_action::none),
+            on_delete(std::ref(*this), false, foreign_key_action::none) {}
+
+        foreign_key_t(const foreign_key_t& other) :
+            _columns(other._columns), _references(other._references),
+            on_update(std::ref(*this), true, other.on_update._action),
+            on_delete(std::ref(*this), false, other.on_delete._action) {}
+
+        foreign_key_t& operator=(const foreign_key_t&) = delete;
+
+#ifdef SQLITE_ORM_DEFAULT_COMPARISONS_SUPPORTED
+        friend bool operator==(const foreign_key_t& lhs, const foreign_key_t& rhs) = default;
+#else
+        friend bool operator==(const foreign_key_t& lhs, const foreign_key_t& rhs) {
+            return lhs._columns == rhs._columns && lhs._references == rhs._references &&
+                   lhs.on_update == rhs.on_update && lhs.on_delete == rhs.on_delete;
+        }
+#endif
+    };
+
+    /**
+     *  Cs can be a class member pointer or column pointer
+     *  Available since SQLite 3.6.19
+     */
+    template<class... Cs>
+    struct foreign_key_intermediate_t {
+        using tuple_type = std::tuple<Cs...>;
+
+        tuple_type _columns;
+
+        /**
+         *  Specify one or more target fields, which can either be pointers to class members or column pointers.
+         */
+        template<class... Rs>
+        foreign_key_t<tuple_type, std::tuple<Rs...>> references(Rs... refs) {
+            return {std::move(_columns), {std::forward<Rs>(refs)...}};
+        }
+
+        /**
+         *  Specify one or more target fields that are member pointers of base classes,
+         *  specifying the derived class as an explicit template argument.
+         */
+        template<class O, class... Base, class... F>
+        foreign_key_t<tuple_type, std::tuple<F O::*...>> references(F Base::*... refs) {
+            static_assert(polyfill::conjunction<is_field_of<F Base::*, O>...>::value,
+                          "Referenced fields must be from explicitly specified derived class");
+            return {std::move(_columns), {refs...}};
+        }
+
+#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
+        /**
+         *  Specify one or more target fields that are member pointers of base classes,
+         *  specifying the derived class as an explicit template argument.
+         */
+        template<orm_table_reference auto table, class... Base, class... F>
+        auto references(F Base::*... refs) {
+            return this->references<auto_decay_table_ref_t<table>>(refs...);
+        }
+#endif
+    };
+#endif
+
+    template<class T>
+    constexpr bool is_foreign_key_v =
+#if SQLITE_VERSION_NUMBER >= 3006019
+        polyfill::is_specialization_of<T, foreign_key_t>::value;
+#else
+        false;
+#endif
+
+    template<class T>
+    struct is_foreign_key : polyfill::bool_constant<is_foreign_key_v<T>> {};
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3006019
+    /**
+     *  FOREIGN KEY constraint factory function taking one or more fields, which can either be pointers to class members or column pointers.
+     *  Available since SQLite 3.6.19
+     */
+    template<class... Cs>
+    constexpr internal::foreign_key_intermediate_t<Cs...> foreign_key(Cs... columns) {
+        return {{std::forward<Cs>(columns)...}};
+    }
+
+    /**
+     *  FOREIGN KEY constraint factory function taking one or more fields that are member pointers of base classes,
+     *  specifying the derived class as an explicit template argument.
+     *  Available since SQLite 3.6.19
+     */
+    template<class O, class... Base, class... F>
+    constexpr internal::foreign_key_intermediate_t<F O::*...> foreign_key(F Base::*... columns) {
+        static_assert(polyfill::conjunction<internal::is_field_of<F Base::*, O>...>::value,
+                      "Fields must be from explicitly specified derived class");
+        return {{columns...}};
+    }
+
+#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
+    /**
+     *  FOREIGN KEY constraint factory function taking one or more fields that are member pointers of base classes,
+     *  specifying the derived class as an explicit template argument.
+     *  Available since SQLite 3.6.19
+     */
+    template<orm_table_reference auto table, class... Base, class... F>
+    constexpr auto foreign_key(F Base::*... columns) {
+        return foreign_key<internal::auto_decay_table_ref_t<table>>(columns...);
+    }
+#endif
+#endif
+}
+
+// #include "schema/constraints/unique.h"
+
+// #include "schema/constraints/check.h"
 
 // #include "../table_info.h"
 
@@ -13264,6 +13275,39 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 // #include "../field_of.h"
 
 // #include "column.h"
+
+namespace sqlite_orm::internal {
+    template<class Elements>
+    using col_index_sequence_of = filter_tuple_sequence_t<Elements, is_column>;
+
+    template<class Elements, class F>
+    using col_index_sequence_with_field_type = filter_tuple_sequence_t<Elements,
+                                                                       check_if_is_type<F>::template fn,
+                                                                       field_type_t,
+                                                                       filter_tuple_sequence_t<Elements, is_column>>;
+
+    template<class Elements, template<class...> class TraitFn>
+    using col_index_sequence_with = filter_tuple_sequence_t<Elements,
+                                                            check_if_has<TraitFn>::template fn,
+                                                            constraints_type_t,
+                                                            filter_tuple_sequence_t<Elements, is_column>>;
+
+    template<class Elements, template<class...> class TraitFn>
+    using col_index_sequence_excluding = filter_tuple_sequence_t<Elements,
+                                                                 check_if_has_not<TraitFn>::template fn,
+                                                                 constraints_type_t,
+                                                                 filter_tuple_sequence_t<Elements, is_column>>;
+
+    template<class Elements>
+    using hidden_col_index_sequence_of = filter_tuple_sequence_t<Elements, is_hidden_column>;
+
+    template<class Elements, class F>
+    using all_col_index_sequence_with_field_type = filter_tuple_sequence_t<
+        Elements,
+        check_if_is_type<F>::template fn,
+        field_type_t,
+        filter_tuple_sequence_t<Elements, mpl::disjunction_fn<is_column, is_hidden_column>::template fn>>;
+}
 
 namespace sqlite_orm::internal {
 
@@ -21187,7 +21231,127 @@ namespace sqlite_orm::internal {
 
 // #include "window_functions.h"
 
-// #include "constraints.h"
+// #include "column_constraints.h"
+
+// #include "table_constraints.h"
+
+// #include "fts5_constraints.h"
+
+// #include "schema/constraints/prefix.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <utility>  //  std::move
+#endif
+
+namespace sqlite_orm::internal {
+    template<class T>
+    struct prefix_t {
+        using value_type = T;
+
+        value_type value;
+    };
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
+    /**
+     *  prefix=N table constraint factory function. Used in FTS virtual tables.
+     *
+     *  https://www.sqlite.org/fts5.html#prefix_indexes
+     */
+    template<class T>
+    constexpr internal::prefix_t<T> prefix(T value) {
+        return {std::move(value)};
+    }
+#endif
+}
+
+// #include "schema/constraints/tokenize.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <utility>  //  std::move
+#endif
+
+namespace sqlite_orm::internal {
+    template<class T>
+    struct tokenize_t {
+        using value_type = T;
+
+        value_type value;
+    };
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
+    /**
+     *  tokenize='...'' table constraint factory function. Used in FTS virtual tables.
+     *
+     *  https://www.sqlite.org/fts5.html#tokenizers
+     */
+    template<class T>
+    constexpr internal::tokenize_t<T> tokenize(T value) {
+        return {std::move(value)};
+    }
+#endif
+}
+
+// #include "schema/constraints/content.h"
+
+#ifndef SQLITE_ORM_IMPORT_STD_MODULE
+#include <utility>  //  std::move
+#endif
+
+// #include "../../alias_traits.h"
+
+namespace sqlite_orm::internal {
+    template<class T>
+    struct content_t {
+        using value_type = T;
+
+        value_type value;
+    };
+
+    template<class T>
+    struct table_content_t {
+        using mapped_type = T;
+    };
+}
+
+SQLITE_ORM_EXPORT namespace sqlite_orm {
+#if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
+    /**
+     *  content='' table constraint factory function. Used in FTS virtual tables.
+     *
+     *  https://www.sqlite.org/fts5.html#contentless_tables
+     */
+    template<class T>
+    constexpr internal::content_t<T> content(T value) {
+        return {std::move(value)};
+    }
+
+    /**
+     *  content='table' table constraint factory function. Used in FTS virtual tables.
+     *
+     *  https://www.sqlite.org/fts5.html#external_content_tables
+     */
+    template<class T>
+    constexpr internal::table_content_t<T> content() {
+        return {};
+    }
+
+#ifdef SQLITE_ORM_WITH_CPP20_ALIASES
+    /**
+     *  content='table' table constraint factory function. Used in FTS virtual tables.
+     *
+     *  https://www.sqlite.org/fts5.html#external_content_tables
+     */
+    template<orm_table_reference auto table>
+    constexpr auto content() {
+        return content<internal::auto_decay_table_ref_t<table>>();
+    }
+#endif
+#endif
+}
 
 // #include "conditions.h"
 
@@ -21964,23 +22128,16 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 #include <concepts>  // std::convertible_to
 #endif
 #include <string>  //  std::string
-#include <tuple>  //  std::tuple_element, std::make_tuple
-#include <utility>  //  std::forward, std::move
+#include <utility>  //  std::move
 #endif
 
 // #include "../functional/cxx_type_traits_polyfill.h"
 
 // #include "../functional/gsl.h"
 
-// #include "../functional/mpl.h"
-
 // #include "../type_traits.h"
 
-// #include "../constraints.h"
-
 // #include "table_base.h"
-
-// #include "column.h"
 
 namespace sqlite_orm::internal {
 #ifdef SQLITE_ORM_CPP20_CONCEPTS_SUPPORTED
@@ -27347,7 +27504,7 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 #include <string>  //  std::string
 #endif
 
-// #include "constraints.h"
+// #include "schema/constraints/default.h"
 
 // #include "serializer_context.h"
 
@@ -27396,13 +27553,9 @@ namespace sqlite_orm::internal {
 #include <algorithm>  //  std::find_if, std::ranges::find
 #endif
 
-// #include "../tuple_helper/tuple_filter.h"
-
 // #include "../type_traits.h"
 
 // #include "../type_printer.h"
-
-// #include "../schema/column.h"
 
 // #include "../schema/table.h"
 
@@ -28486,13 +28639,13 @@ SQLITE_ORM_EXPORT namespace sqlite_orm {
 
 // #include "../functional/mpl.h"
 
-// #include "../field_of.h"
+// #include "../member_traits/member_traits.h"
 
 // #include "../schema/virtual_table.h"
 
 // #include "../schema/column.h"
 
-// #include "../constraints.h"
+// #include "../fts5_constraints.h"
 
 #if SQLITE_VERSION_NUMBER >= 3009000 || defined(SQLITE_ORM_ENABLE_FTS5)
 namespace sqlite_orm::internal {
